requirements.txt
selenium==3.141.0
helium==3.0.7
webdriver-manager==3.4.2
playwright==1.14.1
dagster==0.12.8
dagit==0.12.8
tenacity==8.0.1
mpire==2.2.1
neptune-client==0.10.8
python-dotenv==0.19.0